+++ src/indicators/idmc_displacement/utils/wrangle_displacement.R
@@ -4,6 +4,10 @@ box::use(
   zoo
 )
 
+box::use(
+  cs = src/utils/cloud_storage
+)
+
 #' Wrangle displacement data
 #'
 #' Takes in the IDMC displacement data and turns it into the proper format for generating
@@ -13,6 +17,11 @@ box::use(
 #'     displacement and other types of displacement.
 #' * Generates 7 day and 30 day rolling sums for plotting and alerting respectively.
 #'
+#' If the dataset is on conflict, the IDMC IDU currently has issues where only
+#' very recent years, maybe 2022 or 2023 onward, has reliable data for certain countries.
+#' To fix this, we currently filter the data based on start dates stored in
+#' `input/idmc_dates.parquet`.
+#'
 #' @param df_raw Raw conflict data frame
 #'
 #' @returns Wrangled data frame
@@ -20,17 +29,17 @@ box::use(
 #' @export
 wrangle <- function(df_raw) {
   df_raw |>
-    dplyr$mutate(
-      displacement_type = ifelse(
-        displacement_type == "Conflict",
-        "Conflict",
-        "Disaster"
-      )
-    ) |>
     dplyr$group_by(
       displacement_type, iso3
     ) |>
     idmc$idmc_transform_daily() |>
+    dplyr$left_join(
+      y = cs$read_az_file("input/idmc_dates.parquet"),
+      by = c("iso3", "displacement_type")
+    ) |>
+    dplyr$filter(
+      is.na(start_date) | date >= start_date # filter to consistent time series
+    ) |>
     dplyr$group_by(
       displacement_type, iso3
     ) |>
@@ -38,6 +47,7 @@ wrangle <- function(df_raw) {
       displacement_7d = zoo$rollsumr(displacement_daily, k = 7, fill = NA),
       displacement_30d = zoo$rollsumr(displacement_daily, k = 30, fill = NA)
     ) |>
+    dplyr$ungroup() |>
     dplyr$select(
       iso3,
       displacement_type,
@@ -45,6 +55,5 @@ wrangle <- function(df_raw) {
       displacement_daily,
       displacement_7d,
       displacement_30d
-    ) |>
-    dplyr$ungroup()
+    )
 }